+++ CHANGELOG.md
@@ -2,6 +2,7 @@
   - add Amazon Secrets Manager (ASM) confidant
   - print command errors to stderr
   - provide clearer error message on configuration issues
+  - use bytes rather than string on slashing protection import
 
 # Version 1.1.0
   - garbage collect the slashing database on startup to reduce on-disk size